Don't overwrite null projections with translation
In 32f257 we're also adjusting the projected value for rows which
aren't in every top (note that this is only useful for debugging
since those rows are truncated anyway). The adjustment should borrow
the projected value, not the group key. These two values are not equal
for group by a single column which has null values in it since the
null key is translated when used as a hash key so that we can keep
the buffer linear.