Fix: handle nulls in the catalog
Null entries in the catalog would previously not be handled/included in query
results. This caused entries in the column with nulls to no longer be aligned
with their respective rows. This could further cause bad accesses/segfaults
since the number of entries for that column vector is no longer equal to the
number of rows.