web: add support for multiplexing multiple backends
To use, specify multiple comma-delimited servers to `--backend-url`.

Default is to randomly select a server. Use `--round-robin` to
round-robin between multiple servers.

Client is given a session cookie named `session` on first Thrift call,
which maps the cookie to a specific server.

Note: The master server (used for links) may not be multiplexed at this
time.