Initial new result set code drop
Only iteration and reduction for now, only numeric types. Not
integrated with the executor yet, but we have unit tests. A lot
more layouts supported compared with the current result set: all
combinations of columnar and keyless for both baseline and perfect
hash implementations. Also, arbitrary per column compaction is
supported for all these combinations except for the baseline layout.