Make RuntimeFunctions.cpp use our inlining macros
Depending on LTO options, we might need to turn off always inlining when we
compile with -fPIE / -fPIC (for example for using TSAN). Route this through
our macros so that we can re-define ALWAYS_INLINE to nothing in such cases.
Also, make sure we don't compile with nvcc by accident.