Move retrieving leaf results lower in the stack
When the aggregator tries to finish a Sort node, it'll execute the input of it
through RelAlgExecutor::executeWorkUnit, that's where we need to use the result
from the leaves. RelAlgExecutor::executeRelAlgStep is a level too high since
it doesn't care whether that step is a Sort node, in which case the first part of
its execution must come from the leaves and only the sort happens in the aggregator.