Fix races in baseline reduction
Initialization by the winning thead must be completed before the other threads
using the same slot can proceed. In order to achieve this, we need to use
EMPTY_KEY_64 - 1 as a locked spinlock value, which makes it an invalid value
for big integer. We should audit the places where this matters, it's quite clear
the import validation will need to change, but probably the overflow detection
as well. However, it's extremely unlikely to hit this issue since EMPTY_KEY_64 - 1
is tens of quintillions. Also, fix the lack of initialization for single key group
by when the baseline layout is used, which is exposed when using count distinct.