Dynamically figure out view column types
Remove the last use of column metadata stored in SQLite. Description should be
improved, now it claims it's a table, but at least the column types are correct.
Next, we'll route the view creation through the Calcite-based parser and stop
creating the column metadata for views, no one needs it anymore.