Tempered the aggressiveness of qual short-circuiting
Don't mess with the order too much.
Don't touch simple quals, these are simple comparisons that are yanked out of
the filter and put forward, not something we want to short-circuit.
Do look at remaining quals - split them into primary and deferred as usual.
If encountered a highly unlikely qual, leave it as the last primary qual and
defer the rest of the quals. Preserve the original order.

Should fix #654