Fix multiple bugs in thrift API and mapdql commands related to privileges
I was supposed to fix privileges' reporting for roles, but it turned out much bigger.
Here is what this commit contains:
- Allow get_db_objects_for_role to be actually used on roles. Was working for
users only. Also, renamed it to get_db_objects_for_grantee to make its purpose
clearer.
- Close small security defect: get_db_objects_for_grantee can be used only on
roles visible to the current user (if it's not a superuser).
- Removed several either redundant or even not implemented (no idea why even
add these) Thrift APIs, before anybody starts relying on those.
- Implemented get_db_object_privs Thrift method, which returns whatever privileges
current user has either directly or via roles on the given object.
- Added a second parameter for \\object_privileges command - object type. It's
necessary because it's currently possible to have, for example, a database and a
table with the same name.
- Renamed get_all_roles to just get_roles, for similarity with get_users.
- Also made some small cleanup and refactoring to mapdql, and some other minor
changes