Only use frag offsets for outer table columns
The inner table is required to be 1 fragment / shard, so we don't need
to worry about frag offsets. We do need to apply an offset to rowid, but
that is handled separately in codegenRowId now.

Since RowID is the only col that will ever need to be offset, even in
non-sharded joins, we can simply use a nullptr for all frag offets other
than the offset for the first (outer) table in the join.