Serialize geospatial types as GeoTargetValues
Adds support for serializing a geospatial type as a new top level GeoTargetValue. Serializing as a GeoTargetValue will provide access to the raw data stored in each of the geospatial columns that make up the type.

All geo serialization in the ResulSet was also moved into a header file to separate from the main ResultSetIteration.cpp functionality.