Fix out of bounds col widths index in ResultSet
When reading from the none encoded string buffer attached to
ResultSetStorage, we only need an index into the buffer, and not the
length of the string (since the buffer is a vector of strings). The
QueryMemoryDescriptor will only have one col width for the none enconded
string target in this case, so we need to skip reading the string
length.