Allow e-mails for user names
Extend parser to accept emails. It's supposed to accept any valid e-mail
address in place of a user name and also SOME invalid ones. That is, I rely
here on whatever application uses MapD to check e-mail correctness.
I've tested e-mail in all the privileges commands and in mapdql connect string.
It seems to be working fine.