catalog: explicitly set is_deletedcol when calling addColumn
This allows us to bypass the default value for the `is_deletedcol`
column, which was `false` instead of `0` for a few commits. Data dirs
which were migrated to add the col while the default was `false` would
then end up causing a `bad lexical cast` exception if a column is added.
Since sqlite doesn't support changing the default, solution here is to
just not use the default value.

The `default false` did not make it into a public release, so this
should not affect customers.