Allow NULL literals, cast typeless/valueless NULL constants based on context
Cast first sets the new type and then sets this type's NULL sentinel value.
This is important for array literal initializers containing NULLs, e.g.
in {NULL,1,2,NULL} array literal, 1 and 2 are initially created as SMALLINTs
and NULLs are typeless/valueless constants. If it turns out that what's being
initialized is INTEGER[] then the array literal will be cast to INTEGER[],
i.e. each element will be cast to INTEGER. In this process, NULL constants
will get the type INTEGER and the actual value of NULL_INT sentinel.