Properly fill dict encoded perfect hash table for subqueries
If the outer column is dict encoded text and the result of a subquery, the range of the column is less than or equal to the range of the dictionary. Ignore entries outside this range, since including them will cause the hash table fill routine to exceed its buffer.

(cherry picked from commit 731f9dc655cdb76a0e8b0e27bcfb031a49456839)