Modify distributed locking scheme and introduce TableLockMgr
The TableLockMgr makes the process of obtaining table locks more explicit, requiring the user to manage a vector of read locks and a vector of write locks. While this is increased burden for the caller, we provide structures and several APIs to help manage lock acquisition and lock management. Furthermore, the explicit separation of read/write locks is useful for debugging.