cmake: only link tests to required libs
Experiencing some linking issues when using -std=c++17 where certain
tests would fail to link due to missing symbols, even though the
original library is not actually used in the test. Solution is to only
link those tests to the libraries that they actually use.

Also explicitly linking Boost to `Shared` as it's now required for
Logger.

Curl + dependencies needs to be reworked a bit so that both S3 and GDAL
link to the same place. For now using the "S3 support libraries" var for
GDAL as that var is specifically curl + ssl + crypto.