import: support for line endings in quoted fields for function find_end()
When buffer slicing happens, find_end() function was not able to tell if
the line delimiter is in quoted fields or not. This causes that sliced
buffer ends in the middle of a field even in single thread. Because of
this issue, multi-threading was not working when parsing data which has
quoted fields and threads=1 was being used. Now it supports
multi-threading after this fix.
Automated test is also added.

(cherry picked from commit fd391b2025fc697361500484bf1526bd4d0ca4d1)