Fix boost locale related build failure on PPC
On PPC, build of targets that depend on boost/std locale related code
result in undefined reference errors for references that are defined in
the stdc++ library. This commit directly adds the stdc++ library linker
flag to these targets, when running on PPC, in order to resolve the
build failure.