Clarify representation of permissions for grantee
- objectType was interpreted as representing the DB object type the
privilege was granted on, however it instead was representing the
"permission object type". This is especially confusing since we have two
types of DB level permissions, one that applies to the DB only, and one
that applies to all tables under that DB.
- This commit changes objectType to work as expected. It now represents
the type of the object it is attached to.
- A privilegeObjectType is introduced wich works as objectType did
previously. It represents the type of object this privilege applies to.
- the \privileges command in omnisql has been updated to display the privileges
in a tabular format

Signed-off-by: Alex Baden <alex.baden@mapd.com>