Add WITH option for "explode collections"
* Added option "geo_explode_collections" (per GDAL, geopandas etc.)

When this option is set with in `COPY FROM WITH` or in `CopyParams`, geo import will attempt to split any `MULTI*` geo into multiple rows of the child type. For example, a five-polygon `MULTIPOLYGON` will be split into five separate rows, each containing one of the polygons as a `POLYGON`. All the other fields will be duplicated.

For now, only tables with a single geo column support this feature. For new tables created automatically from a geo file import, this is not a limitation, as the only geo formats we support also only support one geo column. For tables created manually for separate WKT CSV import, there must only be one geo column. In all cases, the geo column type is either automatically created to be, or must be manually created to be, of the child geo type.

If the source file, whether geo or CSV, contains a mixture of `MULTI*` and child type, they will all be imported as the child type. Obviously only the `MULTI*` rows will be split.

This mechanism works with any OGR type, thus providing us with a way to directly import `MULTILINESTRING` or `MULTIPOINT` data.

This feature should be used carefully on source files with many other columns, due to the resulting overhead of duplicating all those columns' data for each resulting row of a `MULTI*` geo.

This feature can only currently be involved from `omnisql` `COPY FROM`, but support can be added to Immerse to set the appropriate flag in `CopyParams`. Note that when this is done, even though an import will result in exploded data, the preview data shown in Immerse before the import will be unexploded.

* Create a reusable CoordinateTransformation per feature group

Do this instead of transforming to the SR which creates one internally for every feature. This will speed up geo import from non-4326 sources.

* Add unit tests for geo_explode_collections

Made the `MULTIPOLYGON`s in the "mpoly" GeoJSONs actually mpolys (with two child polys)
Added a basic "poly" GeoJSON test
Added row-count verifications where missing on several other geo import tests