startomnisci: use https for backend-url if enabled in config file
Previously startomnisci would always use http for the web server's
`backend-url` parameter, even if the config file was set to encrypt
omnisci_server's http endpoint. This ended up breaking encrypted
installs running in Docker (only place where startomnisci is used).

Workaround here is to grep the config file (if provided) for the two
options which enable encryption and, if found, use https for the
`backend-url`.

This will again break if you pass the `--ssl-cert` and
`--ssl-private-key` to `startomnisci`.

Signed-off-by: Alex Baden <alex.baden@omnisci.com>