cmake: add dummy targets to allow concurrent gen jobs
* cmake: add dummy targets to allow concurrent gen jobs

Previously if a lib included the output of a few add_custom_command
jobs, cmake would run each of those jobs serially to generate that
output.

This adds a dummy target for a few of those, which allows the jobs to
run in parallel with the rest of the build.

Shaves off about 90 seconds from a cuda+render build on my machine.

* cmake: don't build generate_cert every build