Fix arrow ipc when no dict in projection
If a projection was executed without a dict result,
the server would hang. This adds a guard to ensure we
are only taking the dictionary path when we need to.

A regression test with only scalar values has also been added.

(cherry picked from commit 52cd0644606af05b4d709408cad5542753b149a0)