Add SHOW SERVERS command for FSI
Adds Calcite and C++ support for the SHOW SERVERS command.
The SHOW SERVERS command is parsed along with an optional
WHERE clause.
The where clause is validated and then used to query SQLite
in Catalog::selectForeignServersForUser.
The output is put into a TQueryResult object and returned.
This also adds a column for 'creation time' to the catalog
table for foreign servers, which is part of the output.