cmake: begin linking with blosc again
It turns out there are some issues with how we're building
parquet/arrow, where the zstd lib isn't getting pulled in correctly.
This hadn't been noticed in the past because the missing symbols were
instead getting pulled in from blosc. Fix for now is to start linking
against blosc again.