CSV data wrapper append support
This makes the CSV Data wrapper aware of 'APPEND' update mode.
When in append mode the Data Wrapper will not parse any previously parsed data, and instead just load any new data.
This does not include support in the FSI cache to prevent the eviction of data, that will be coming in a later PR.
I haven't added support for S3 yet as we need to setup a new S3 bucket to test it, but it will be straightforward to implement based on the local multifile reader.

It also fixes an issue with multi-file foreign tables with missing newlines, and adds test cases.
It moves the function add_end_of_line_if_needed so it is applied by CsvReader rather than the ArchiveReader so it can be adjusted at the end of any files, instead of when CsvReader doesn't fill the entire buffer (which could fail if an EOF coincided with the end of a buffer).