Added missing resetChunkStats() implementations for Array types.
Array encoders had no implementation for resetChunkStats().  This was causing problems in the Disk Cache because it relied on resetChunkStats() to write metadata statistics.  As a result, Arrays were not having their statistics written to the cache's metadata and this would cause us to read unitialized values.  Intermittently this would cause an error if the uninitialized values read into the min/max values would give us invalid combinations that would cause checks to fail (if max < min && min < 0).

This change adds the missing implementations and adds additional testing to cover the missing types.