Reworked option verification for foreign tables and servers
Reworked option verification for foreign tables and servers

This item changes foreign table/server verification to require only one of file_path & base_path to be set. Prior to this base_path was always required.
In addition to the feature change, there is some additional comments and refactoring involved (a few verification checks were being replicated in multiple places) and some function name changes to clarify behaviour.
The getFilePath() section changes are a little more than might be expected because they are merging the function that was already there with the implemenation that was in the ParquetDataWrapper.