Removed disk cache size limit and eviction algorithm
* Removed disk cache size limit and eviction algorithm

Due to optimizations in the ForeignStorageMgr, the ForeignDataWrappers are able to write directly to the FileBuffers that underly the ForeignStorageCache. This means that the cache is unable to properly enforce any size limitations, as the FileBuffers will allocate more disk space as necessary to complete the DataWrappers' requests. Currently the Cache is enforcing the number of FileMgr pages that are in use, but not the allocated space, making the feature misleading at best.

In addition, the FileMgr currently does not provide any interface to defragment and remove files once they have been allocated, so there is no way to shrink the cache's allocated space without evicting at the table level.

Considering those options, we have decided to remove the cache size limitation feature from the 5.5 release, with intentions on bringing it back in a future release when the proper infastrucure can be added.

* Fixed show cache detail tests to remove cache size limit column