FSI CSV - Only convert to data blocks once per thread
When multiple file regions are parsed per thread on CSV chunk fetch, the import buffer to datablock conversion is called multiple times. This is functionally ok for most types, but for string arrays the previous data is not cleared and we end up with repeated values.
This change only runs the conversion after the final file region is parsed. A test was extended to expose this case.