cmake: java: ignore changes in generated files
Previously running `cmake` after an initial build would cause all *.java
files, including generated ones, to get listed as dependencies for the
`mapd_java_components` target. This is a problem because that target is
what generates some of those files in the first place. This in turn lead
to problems where running `make` after a `make clean-all` would fail due
to missing dependencies.

Workaround is to remove generated files from the list of dependencies.
This previously was not possible due to the previous minimum supported
version of cmake, which did not support `file(filter)`.