Add a workaround for file renaming problem.
On Windows we cannot rename directory if there are
some files still open. In practice we see that even
if we close all files in a single thread and then
instantly try to rename a directory, we still get
an error. An ugly workaround for this problem is to
wait for some time and try again. The issue requires
additional investigation.