S3 Select client and tests
This class will be used by the S3 select based CSV wrapper, allowing access to individual columns, as well as metadata generation for some types.
The tests are currently not run as part of sanity_tests as a valid AWS_ACCESS_KEY_ID needs to be set in the environment to run the test.