Injecting threshold into all distance functions
If distance dips below the threshold, function will return whatever it has.
ST_DWithin functions will pass dwithin distance as threshold to the underlying
ST_Distance which will bail as soon as geometries are determined to have gotten
close enough to each other. Actual distance, if it's smaller, is not important.