Arrow 2.0.0 Support and folly upgrade
* Update arrow dep to 2.0.0

* Update for Arrow 2.0.0 dictionary changes

* format

* Add back num field check

* Upgrade folly to v2021.02.01.00

* Fix longstanding issues with Arrow tests

This addresses the following:

1) Dates can be returned as date32 if we're serializing arrow for CPU
and date64 if were serializing for GPU. The ArrowResultSetConverter
constructor we were using in the tests defaulted to GPU, and we never
tested the CPU path.

2) Switch to using Arrow Wire when testing Arrow serialization for the
tests, rather than the custom test routine we were using before. This
ensures parity b/t what is used in production and reduces duplication.

3) Fix an ownership issue in the Arrow Wire code that was detected in
the tests after switching to use it for serialization. The lifetime of
the allocated buffer was not correct, and the buffer was silently being
destroyed in some cases, which was not caught and leading to a corrputed
buffer.

* use install_folly in centos & ubuntu1604 scripts

* cmake: arrow: aws: add new aws libs required by arrow 2.0.0

* Remove vendored fmt lib

* Fix linking issue when using LLD

* Need libevent dep on centos

* Revert "Need libevent dep on centos"

* Add libevent dep needed for folly

* deps: add -fPIC for bzip2, xz, libarchive

* Fix encoder test linkage

* Remove superfluous log message

* Fix dropped lib

Co-authored-by: Andrew Seidl <dev@aas.io>
(cherry picked from commit 4c382c98c4334db8a83f62a4de6e11f01bf35a30)