cmake: thrift: remove _constants.cpp files from list of generated thrift files
Thrift 0.14 no longer generates the _constants.cpp files, which caused
cmake to complain about missing files. Deps will
likely be moved to this version soon, but in the meantime the workaround
is the simply no longer list the _constants.cpp files as required. This
should be fine because the files do not get modified by hand and older
versions of thrift will always generated the _types.cpp and
_constants.cpp files, so we really only need to list the _types.cpp as
required.