added table-specific checkpoint to CachingFileMgr
* Added table-specific checkpoint to CachingFileMgr

Caching file mgr was experiencing tsan issues because it used a global checkpoint while FileBuffers from multiple tables were being written to. This changeset adds a table-level checkpoint function and supporting functions to checkpoint only table-specific FileBuffers.

In addition, this change needed to add an epoch lock since the FileBuffers from one table may query which epoch the Mgr is on while a checkpoint may be invoked by another table (which increments the epoch).

There is a further issue with these epochs being incremented while buffers are being written, but this will be addressed in a subsequent PR that splits the epochs by table. This should also remove the need for these locks.