Handle DOS carriage return in s3 select
Since its possible for null entries to be returned as empty lines from S3 Select calls, the code that skipps empty lines when parsing csv files was turned off for S3 Select. The also disabled handling of DOS style carriage return line feed endings ("\r\n").
These can be returned by S3 Select calls because the \r is treated as a part of the last field.
This change allows a single "\r\n" to be treated as a single line ending for s3 select.