limit optional buffers
There is an issue in FSI where we would attempt to aggressively prefetch optional buffers regardless of the disk cache or memory constraints of the system we are on. This was leading to a case where the system had enough space to perform a query, but the added memory needed to simultaneously prefetch the additional buffers was causing us to crash (symptoms depended on whether the disk cache was enabled or not).
The solution in this PR is to limit the number of optional buffers we attempt to prefetch based on the maxChunkSize of the table (maxChunkSize is not currently set correctly for foreign tables, but the default value is still usable so long as we aren't creating massively large chunks).
If the limit is hit we will prioritize chunks in the same fragment before attempting to prefetch chunks in other fragments.