Use assigned GPU device id in memory system table tests
The GPU device id that is assigned to a fragment depends on both the
table id and fragment id when the g_use_table_device_offset flag is
set. Update the assertion in the GPU memory system table tests to use
assigned device id and not assume assignment to the first GPU device.