cmake: java: place java test logs into build dir
Some of the java tests run automatically during the build, which it
turns out to be somewhat problematic because the new version of log4j
will now successfully write its logs even if the variable the log path
is based on is not defined. This means the logs get written to the
source tree rather than the build tree.

Simple solution is to pass in the log path variable when running mvn
install.