Compressed ST_Intersects translation support for reverse arg orders
Proactively try to compress ST_Intersects's first argument to preempt
possible argument swap in a switch to compressed implementation.

Swap ST_Intersects arguments if that allows a switch to cIntersects.

ST_Disjoint is a negated ST_Intersects, so it follows the same optimizations.

Removed redundant cIntersects extension functions which now go unused
(and also not supported by the Overlaps framework).