Enforce Raster Column/Band Type Match
When Raster Import is implemented in Immerse, the intention for now is to disallow column type overriding. This may be relaxed in the future, but for now this PR enforces that, if the table is created in advance by Immerse, that the column types are still what was detected, oherwise an exception is thrown.