Let CachingFileMgr recover from reading uncheckpointed pages
When a table is dropped, the CachingFileMgr will delete the table metadata and free the associated table pages. Freeing a page does not mean it is deleted, however, just that it is added to the in-memory list of available pages.

This means that on server restart, when the cache attempts to repopulate from disk, we can end up reading pages on disk that were "deleted". When this occurs we check if the page is valid by comparing the page's epoch to that of the FileMgr (or in the CachingFileMgr's case, the epoch it records for that table). If the CFM has no metadata for that table, it will currently crash as it expects us to only request the epoch for a table it is aware of.

This PR changes the CFM::epoch() function to automatically return a minimum epoch for tables that the CFM has no metadata for, so that this comparison against deleted pages can now be executed correctly leading to the page being deleted due to it's advanced epoch, rather than a crash.