enable distributed system tables
Enabled system tables in distributed mode and adjusted tests to accommodate.

There are still some gaps that need to be addressed in follow-up items:

As system tables are not replicated and do not have a shard_key value they can't be joined on a dictionary column.
The PR has now been updated with subtasks that support the SHOW DISK CACHE USAGE command and system dashboards in distributed. The latter requires some parsing of the query in the DistributedHandler which is not ideal, but should suffice to get this started.

There are some issues with testing system tables in distributed because our testing runs all the distributed tests on the same cluster while the aggregator is reset between tests (technically this is an unsupported case according to our documentation). This is causing some issues with the TablesSystemTables test (likely because it is using temporary dictionaries), so it has been temporarily disabled in distributed. The test works fine if you run each test after restarting the cluster (as our documentation requires).