deps: bump to gdal 3.4.1, proj 8.2.1, enable netcdf, hdf5, and blosc
* deps: rework thrift to only build compiler + c++ lib

By default it tries to build the libs for any language which the system
has dev dependencies for, which often ends up breaking stuff.

Note that we have to explicitly disable go and python because they don't
seem to respect the enable-libs flag.

Also moving to the common-functions script so that it only has to be
maintained in one place.

* deps: move blosc to common-functions.sh and bump to 1.21.1

* deps: add netcdf

* deps: bump to gdal 3.4.1, proj 8.2.1, enable netcdf and blosc

* Fix netCDF build

Co-authored-by: Andrew Seidl <dev@aas.io>
Co-authored-by: Simon Eves <simon.eves@heavy.ai>