Enabled system tables by default
Enable system tables by default.

Changes:

Enabled system tables by default. This caused some tests in CatalogMigrationTest to fail.
Fix CatalogMigrationTest failures. The issue was that the test make some assumptions as to what was enabled by default, so we had cases where system tables were enabled while FSI was disabled (not a legal case).
Add new pre-condition/post-condition checks. Added some checks that guard against the above case in the catalog (make sure the options that require FSI can't be turned on without FSI), and some checks to make sure data exists after a function fetches it.
Code outlining and refactor. Outlined a bunch of code in Catalog::buildMaps() to build each map in a new function. This was already what was being done for some maps but now it is more consistent and legible.
Updated code format in refactor. Use auto more aggressively, use std namespace where appropriate, etc...
Code reordering. Hoisted createDefaultServersIfNotExists() out of buildMaps() since it is not a part of the map building process. This technically means we execute it a bit later in the initialization (after buildMaps() instead of in the middle of) but this should not change the semantics as it only alters the maps if there was no entry in the catalog.
Renamed some functions such as addForeignTableDetails() to better reflect their purpose (they update existing map entries).

(cherry picked from commit 5962b638525911bc33d3a89ff6c7d47fb2583c52)