Fix geo file import thread contention
Move OGCT construction to parent thread

Even with the comparison OGRSpatialReference objects now being per-thread, it's possible that multiple OGRFeatures all refer to a single OGRSpatialReference in the data set, hence there could still be contention on the comparison.

Move that entire block to the parent thread, and just pass any resulting OGRCoordinateTransformation* into the worker.

The correct structure is to open a data set per thread, then everything is thread-safe. We must get this right for the FSI port.

Co-authored-by: Simon Eves <simon.eves@heavy.ai>