odbc order by as separate option
Added a new SQL_ORDER_BY option for odbc foreign tables. This option splits out the order by clause that was implicitly required in the SQL_SELECT option to make sure that the user is aware they need to specify the result order.
Refactored some of the code in odbc_utils to use an OdbcSelectDescriptor instead of passing the same 6 parameters to multiple functions.
Added a warning when accessing legacy odbc foreign tables that they need to add an ORDER_BY command using the ALTER FOREIGN TABLE command (support added).
Fixed a bug where clang-format was disabled for a portion of odbc_utile.cpp.
Updated importer and copy-params to support the new odbc table syntax.
Updated tests to to support the new syntax.
Updated tests in ForeignTableDmlTest to call GTEST_SKIP() earlier and properly avoid setup that should be skipped (earlier they were still initializing things like odbc servers even when odbc was disabled). Also updated the TearDown() methods so that they can be called without assuming that SetUp() was called (in case it was skipped).