Renamed data files before attempting to alter them
Fixed two issues with dump/restore if the table was dumped pre-rebranding and restored post-rebranding.

The first issue is that the adjust_altered_table_file() function in the TableArchiver is being called before we rename the data files to their new names, and so the above function doesn't recognize the data files and does not adjust them correctly. This means we have no mapping for adjusted columns so the new table is unaware of any deleted or altered columns.

The second issue is that the TableArchiver is crashing if it runs into a page with column information for a column that was removed (even if the page is marked for deletion because the column was dropped). This is an inconsistency in how the FileMgr and TableArchiver parse the data files.

The fix for the first issue included here is to alter the data file names before attempting to alter their contents.

The fix for the second issue is to

Some additional changes:

Changed the renaming code so that we aren't altering filenames as we are iterating over them.
Added an placeholder entry in DatumToString() for geo types so that we don't crash every time we try to dump metadata that contains a geotype.
Changed the FileMgr::hasChunkMetadataForKeyPrefix(). Not super relevant to this fix, but debugging took me through it and it needed updating from previous changes to the disk cache system.
Split up the adjust_altered_table_files() function so that we aren't creating a lambda in a heavily nested scope (makes reading easier).
Clarified some of the data file reading code with variable names (instead of using a buffer array and arbitrary indexes).

(cherry picked from commit a0402a910de7770568175108d59cb93640aa3ce0)