Fix ForeignTableRefreshConcurrency Test for Distributed.
Changed select query for ForeignTableRefreshConcurrencyTest to one that will touch both tables while also being supported in distributed mode. Previously we executed a join that is not supported in distributed because the tables are not replicated. This has been changed to two select statements to make sure we still touch both tables.