Fix hang that can occur for logs system tables and set a cap on the number of processed log files
* Fix hang that can occur for logs system tables and other foreign
tables that are based on data wrappers that inherit from
AbstractTextFileDataWrapper. The hang was caused by a bug in
AbstractTextFileDataWrapper where ParseBufferRequest objects were not
added back to the request pool when request processing was skipped for
empty buffers that were a result of reading empty files.

* Add a server option that specifies the maximum number of log files
that should be processed by each logs system table.