Change ExpressionRange calculation to recognize placeholder metadata as an invalid range.
Currently the HeavyConnect (FSI) infrastructure will skip the calculate of metadata on expensive chunks so that they can be evaluated on demand (lazily). For such "skipped" chunks it inserts "placeholder metadata" which records the number of rows in the chunk, type, and other known metadata - but uses default metadata for entries like min and max (default being min > max, usually min = max and max = lowest).

Currently the ExpressionRange calculation assumes that any chunk with min > max is empty, as this is the value set on a chunk when it is created with no entries. This causes problems, particularly with joins using a perfect-hash because we assume certain fragments are empty based on this expression range, when they are actually supposed to represent "unknown", or "uncalculated" metadata.

There are three salient changes in this PR (as well as some refactoring to make debugging easier).

Placeholder metadata has been defined as (min > max && numElements > 0 && has_nulls == false). This change was needed because this is a metadata state that is otherwise not legal and therefore there is no confusion between a chunk that "happens" to have these metadata and a chunk that has placeholder metadata. Prior to this change, placeholder metadata could have has_null == true and therefore was indistinguishable from a chunk that had only null values (it would have the same metadata). This PR experimented with having ChunkMetadata record placeholder metadata separately, but that proved to be untenable because we currently have no system for storing that additional field on disk (ChunkMetadata are not persisted - they are created from Encoders, which are persisted to disk).

The ParquetDataWrapper now returns placeholder metadata, but internalizes a more accurate has_nulls value. This was needed because the metadata created by ParquetDataWrapper::populateChunkMetadata() needs to be recognizable as placeholder metadata when necessary, but the wrapper itself may have access to more accurate information and needs to retain that as it will be built on during chunk population.

ExpressionRange now identifies placeholder metadata and creates an Invalid range with it. This allows us to avoid making incorrect assumptions about placeholder metadata and treat it defensively until we can fetch the data proper and optimize.

Additional changes in this PR are to standardize the creation of placeholder metadata in all datawrappers (some wrappers duplicated code), add more <<() overloads to view objects, move more code out of header files, and to clean up some other code run into while debugging.

(cherry picked from commit b2c0ac98dfeeaed49c427fc786a1b7bd0cb93034)