fix-disappearing-db
This PR fixes an issue with dropping users who own DBs.

The problem manifests in two ways:
1: Once a user is dropped, DBs it owned now longer show up in the show databases command (even when run by a super-use)
2: The DB's owner id was never cleared, so if the user id was re-used when creating a new user, they could automatically be assigned ownership of a pre-existing db.

The solution we opted for here was to prevent a user from being dropped if it currently owns a DB, therefore requiring that the DB be dropped first. This is the same approach PostgreSQL uses.

While this solves the above two issues, there is still a potential issue with the the user owning tables, or other objects, but that is a more in depth issue that will need to be solved later.

The PR also contains some test cleanup to make sure we are now dropping DBs before their owning users.

Also, the show databases functionality was updated so that a super-user can view all dbs regardless of whether or not they are owned (this was tested, but the tests were removed as it should no longer be possible to enter such a state).